+++ mayan/apps/document_states/forms/workflow_template_state_forms.py
@@ -73,5 +73,5 @@ class WorkflowTemplateStateEscalationForm(forms.ModelForm):
 
 class WorkflowTemplateStateForm(forms.ModelForm):
     class Meta:
-        fields = ('initial', 'label', 'completion')
+        fields = ('initial', 'final', 'label', 'completion')
         model = WorkflowState